refactor(playthrough): expose ProfilePlaythrough via the model layer
Move the ProfilePlaythrough type into entities/playthrough/model/types and
re-export it from the model barrel (the .server query keeps a type-only
re-export for its api-barrel consumers). The profile widget's .type.ts now
sources the domain type from model/ instead of an entity .server module, per
the entities-layer rule. Addresses CodeRabbit #14 (PR #361).
