+++ savepoint-tanstack/src/entities/playthrough/api/get-profile-playthroughs.server.ts
@@ -1,21 +1,8 @@
 import { prisma } from "@/shared/lib/db.server";
 
-import type {
-  PlaythroughKind,
-  PlaythroughStatus,
-} from "../../../../shared/lib/prisma/client.ts";
+import type { ProfilePlaythrough } from "../model/types";
 
-export type ProfilePlaythrough = {
-  id: string;
-  kind: PlaythroughKind;
-  status: PlaythroughStatus;
-  platform: string | null;
-  startedAt: Date | null;
-  finishedAt: Date | null;
-  rating: number | null;
-  notes: string | null;
-  game: { title: string; slug: string; coverImage: string | null };
-};
+export type { ProfilePlaythrough };
 
 /**
  * Recent playthroughs for a user's public profile.
