+++ savepoint-tanstack/src/entities/playthrough/model/index.ts
@@ -3,4 +3,4 @@ export type { AggregatedPlaythroughs } from "./aggregate";
 export { isOrdinalUniqueConflict } from "./ordinal-conflict";
 export { RUN_STATUS } from "./run-status";
 export type { RunStatusEntry } from "./run-status";
-export type { PlaythroughWithEntries } from "./types";
+export type { PlaythroughWithEntries, ProfilePlaythrough } from "./types";
