+++ savepoint-tanstack/src/widgets/profile-overview/ui/profile-playthroughs-section/profile-playthroughs-section.type.ts
@@ -1,4 +1,4 @@
-import type { ProfilePlaythrough } from "@/entities/playthrough/api/get-profile-playthroughs.server";
+import type { ProfilePlaythrough } from "@/entities/playthrough/model";
 
 export type ProfilePlaythroughsSectionProps = {
   playthroughs: ProfilePlaythrough[];
